Inline tutorials list to documentation page
